src/main/java/com/amazon/opendistroforelasticsearch/security/rolesinfo/RolesInfoRequest.java
package com.amazon.opendistroforelasticsearch.security.rolesinfo;

import org.elasticsearch.action.ActionRequest;
import org.elasticsearch.action.ActionRequestValidationException;
import org.elasticsearch.common.io.stream.StreamInput;
import org.elasticsearch.common.io.stream.StreamOutput;
import org.elasticsearch.common.xcontent.ToXContentObject;
import org.elasticsearch.common.xcontent.XContentBuilder;

import java.io.IOException;

/**
 * Used by TransportRolesInfoAction in opendistro_security plugin.
 */
public class RolesInfoRequest extends ActionRequest implements ToXContentObject {

    public RolesInfoRequest() {
        super();
    }

    public RolesInfoRequest(StreamInput in) throws IOException {
        super(in);
    }

    @Override
    public ActionRequestValidationException validate() {
        return null;
    }

    @Override
    public void writeTo(StreamOutput out) throws IOException {
        super.writeTo(out);
    }

    @Override
    public final XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
        builder.startObject();
        builder.endObject();
        return builder;
    }
}
